refactor(SIO): Minor cleanup
