+++ include/hw/dma/apple_sio.h
@@ -3,62 +3,13 @@
 
 #include "qemu/osdep.h"
 #include "hw/arm/apple-silicon/dtb.h"
-#include "hw/misc/apple-silicon/a7iop/rtkit.h"
-#include "system/dma.h"
+#include "hw/misc/apple-silicon/a7iop/base.h"
+#include "hw/sysbus.h"
 
 #define TYPE_APPLE_SIO "apple.sio"
 OBJECT_DECLARE_TYPE(AppleSIOState, AppleSIOClass, APPLE_SIO)
 
-#define SIO_NUM_EPS (0xDB)
-
-typedef struct {
-    uint32_t xfer;
-    uint32_t timeout;
-    uint32_t fifo;
-    uint32_t trigger;
-    uint32_t depth;
-    uint64_t unk18;
-} QEMU_PACKED SIODMAConfig;
-
-typedef struct {
-    uint64_t addr;
-    uint32_t len;
-} QEMU_PACKED SIODMASegment;
-
-typedef struct AppleSIODMAEndpoint {
-    SIODMAConfig config;
-    SIODMASegment *segments;
-    QEMUSGList sgl;
-    QEMUIOVector iov;
-    uint32_t count;
-    uint32_t actual_length;
-    uint32_t id;
-    uint32_t tag;
-    bool mapped;
-    DMADirection dir;
-} AppleSIODMAEndpoint;
-
-struct AppleSIOClass {
-    /*< private >*/
-    AppleRTKitClass base_class;
-
-    /*< public >*/
-    DeviceRealize parent_realize;
-    ResettablePhases parent_reset;
-};
-
-struct AppleSIOState {
-    /*< private >*/
-    AppleRTKit parent_obj;
-
-    /*< public >*/
-    MemoryRegion ascv2_iomem;
-    MemoryRegion *dma_mr;
-    AddressSpace dma_as;
-
-    AppleSIODMAEndpoint eps[SIO_NUM_EPS];
-    uint32_t params[0x100];
-};
+typedef struct AppleSIODMAEndpoint AppleSIODMAEndpoint;
 
 int apple_sio_dma_read(AppleSIODMAEndpoint *ep, void *buffer, size_t len);
 int apple_sio_dma_write(AppleSIODMAEndpoint *ep, void *buffer, size_t len);
